+++ release-notes.md
@@ -1,4 +1,5 @@
 Changes since v2.2.6 include:
+- fix inadvertantly broken 'waiting' app (that's needed by feeder update)
 - UI: improve SDR page for small displays
 - UI: dark mode improvements
 - chart: support shift-click to remove a microfeeder from the chart when all microdfeeders are shown (default)